feat: запретить удаление статуса, если он используется в задачах

+++ app/Services/TaskStatusDeleter.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace App\Services;
 
+use App\Exceptions\TaskStatusIsUsedException;
 use App\Models\TaskStatus;
 use App\Repositories\TaskStatusRepository;
 
@@ -15,6 +16,10 @@ class TaskStatusDeleter
 
     public function delete(TaskStatus $taskStatus): void
     {
+        if ($this->taskStatusRepository->isTaskStatusUsed($taskStatus)) {
+            throw new TaskStatusIsUsedException;
+        }
+
         $this->taskStatusRepository->delete($taskStatus);
     }
 }
